resume-optimizer-server/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.25.1
pdfplumber==0.7.4
openai==0.28.1
httpx==0.23.3
beautifulsoup4==4.12.2
selenium==4.1.0
webdriver-manager==3.5.2
reportlab==4.0.8
python-docx==1.0.1
supabase==1.0.3
waitress==2.1.2
loguru
sentry-sdk
blinker
uvicorn
fastapi
pydantic